Lab-02-Podcast/requirements.txt
azure-ai-projects==2.0.0b2
azure-identity
python-dotenv
openai
